[fix]: merge_results stores the full float32 output
The TP-level merge wrote the final output with avx512_32xfp32_to_32xbf16:
32 bf16 = 64 bytes = only 16 of the float32 lanes in the output buffer,
so HALF the buffer stayed uninitialized - the run-to-run garbage lanes,
the varying medians, and the 192-lane set on the real MiniMax layer-0.
The kernels, the strip packing, and the weights were all proven exact in
isolation (0.0000 rel-err at the production dims, real GGUF data); the
fault was the output store. The merge now writes the float32 values
(the production convention), and the suite's forward default moves to
float32 accordingly.

Also adds the debug_ram_stats ASCII-art RAM map (per-attribute packed
fill, scale stats, down-stage internals; bound as debug_ram_stats and
forwarded through the TP container) for the runtime RAM view.

+++ kt-kernel/test/per_commit/test_moe_gguf_amxint8_cache.py
@@ -196,7 +196,7 @@ def load_cache_cfg(pool, layer_idx, cache_dir):
     return cfg
 
 
-def forward(pool, moe, map_t, seed, out_dtype=torch.bfloat16):
+def forward(pool, moe, map_t, seed, out_dtype=torch.float32):
     torch.manual_seed(seed)
     bsz = torch.tensor([1], dtype=torch.int32)
     expert_ids = torch.stack([torch.randperm(EXPERT_NUM)[:NUM_EXPERTS_PER_TOK] for _ in range(1)]).contiguous()
